+++ resources/js/Pages/DefaultPreset.vue
@@ -51,8 +51,14 @@
 import { ref, onMounted } from 'vue';
 import { Head, router } from '@statamic/cms/inertia';
 import { Field, Checkbox, Button } from '@statamic/cms/ui';
-import axios from 'axios';
 import QueryBuilder from '../components/fieldtypes/QueryBuilder.vue';
+import {
+    FIELD_GROUPS,
+    buildGroupedFields,
+    fetchProductAttributes,
+    fetchSortingOptions,
+    normalizeQuery
+} from '../helpers/productQueryBuilder';
 
 const { enabled: initialEnabled, query: initialQuery, saveUrl } = defineProps({
     enabled: {
@@ -69,107 +75,21 @@ const { enabled: initialEnabled, query: initialQuery, saveUrl } = defineProps({
     }
 });
 
-const normalizeQuery = (query) => {
-    if (!query || typeof query !== 'object') {
-        return { groups: [{ conjunction: 'AND', conditions: [] }], globalConjunction: 'AND' };
-    }
-    const groups = Array.isArray(query.groups) && query.groups.length > 0
-        ? query.groups
-        : [{ conjunction: 'AND', conditions: [] }];
-    return {
-        groups,
-        globalConjunction: query.globalConjunction || 'AND'
-    };
-};
-
 const enabled = ref(initialEnabled);
 const queryValue = ref(normalizeQuery(initialQuery));
-const attributes = ref([]);
 const groupedFields = ref([]);
 const sortingOptions = ref([{ label: 'Created', value: 'created_at' }]);
 const saving = ref(false);
 const attributesLoaded = ref(false);
 
-const fieldGroups = [
-    {
-        key: 'attribute',
-        label: 'Attributes',
-        prefix: 'attribute.'
-    },
-    {
-        key: 'stock',
-        label: 'Stock',
-        prefix: '',
-        fields: [
-            {
-                label: 'Stock Status',
-                value: 'stock_status',
-                type: 'select',
-                operators: ['=', '!='],
-                options: [
-                    { label: 'In Stock', value: 'in_stock' },
-                    { label: 'Out of Stock', value: 'out_of_stock' }
-                ]
-            }
-        ]
-    }
-];
-
-const mapAttributeType = (frontendInput) => {
-    const typeMap = {
-        'text': 'text',
-        'textarea': 'text',
-        'select': 'select',
-        'multiselect': 'select',
-        'boolean': 'select',
-        'price': 'number',
-        'weight': 'number',
-        'date': 'date',
-        'datetime': 'date',
-        'timestamp': 'date'
-    };
-    return typeMap[frontendInput] || 'text';
-};
-
-const mapAttributeOptions = (options) => {
-    if (!options || !options.length) {
-        return [];
-    }
-    const opts = typeof options === 'object' && !Array.isArray(options) ? Object.values(options) : options;
-    return opts.map((option) => ({
-        label: option.store_value || option.option_id,
-        value: option.store_value || option.option_id
-    }));
-};
-
-const buildGroupedFields = () => {
-    groupedFields.value = fieldGroups.map((group) => {
-        if (group.key === 'attribute') {
-            return {
-                label: group.label,
-                options: attributes.value
-            };
-        }
-        return {
-            label: group.label,
-            options: group.fields.map((field) => ({
-                ...field,
-                value: group.prefix ? `${group.prefix}${field.value}` : field.value
-            }))
-        };
-    });
-};
-
-const fetchAttributes = async () => {
+const loadAttributesAndSorting = async () => {
     try {
-        const response = await axios.get('/cp/rapidez/product-attributes');
-        attributes.value = response.data.map((attr) => ({
-            label: attr.frontend_label ? (attr.frontend_label + ` (${attr.attribute_code})`) : attr.attribute_code,
-            value: `attribute.${attr.attribute_code}`,
-            type: mapAttributeType(attr.input),
-            options: mapAttributeOptions(attr.attribute_options)
-        }));
-        buildGroupedFields();
+        const [attributes, options] = await Promise.all([
+            fetchProductAttributes(),
+            fetchSortingOptions()
+        ]);
+        groupedFields.value = buildGroupedFields(attributes, FIELD_GROUPS);
+        sortingOptions.value = options.length > 0 ? options : [{ label: 'Created', value: 'created_at' }];
     } catch (error) {
         console.error('Error fetching attributes:', error);
     } finally {
@@ -201,6 +121,6 @@ const save = () => {
 };
 
 onMounted(() => {
-    fetchAttributes();
+    loadAttributesAndSorting();
 });
 </script>
